requirements.txt
# Pytorch utils for GPU acceleration
--extra-index-url https://download.pytorch.org/whl/cu118
torch
torchvision
torchaudio

# the openai whisper library
openai-whisper
openai

#pandas for dataframe manipulation
pandas

# ipypkernel for VScode
ipykernel

# streamlit for user interface
streamlit
streamlit-audiorecorder

# Azure for the data store
azure-storage-blob

# for login email validation
email-validator